requirements.txt
langchain
langchain-community
langchain-openai
beautifulsoup4
requests
python-dotenv
faiss-cpu
openai